Исправил claims

+++ internal/app/provider/db/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type UserExistResponse struct {
-	UserID     int    `db:"user_id"`
+	UserID     int64  `db:"user_id"`
 	Phone      string `db:"phone"`
 	Email      string `db:"email"`
 	Password   string `db:"password"`
